+++ NEWS.md
@@ -2,6 +2,9 @@
 ## Added
 * Added thumbnail picture for selected sample in "Get signature attributions" tab.
 
+## Changed
+* Made it default to round the reconstructed catalog for plotting.
+
 ## Fixed
 * Fixed a bug when user uploads a VCF with "unknown" region and proceed to do signature 
 attribution.